% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shader-package.R
\docType{package}
\name{shader-package}
\alias{shader}
\alias{shader-package}
\title{shader: Compile and Execute Shaders from R}
\description{
Provides the ability to create, compile, and execute shaders from
    within R. The main focus is on the GLSL language due to its cross-platform
    support. The 'shader' package provides native support for passing in R 
    objects as input and uniforms and render to either a window or to a raster 
    in a headless manner.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/shader}
  \item Report bugs at \url{https://github.com/thomasp85/shader/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (0000-0002-5147-4711)

}
\keyword{internal}
